requirements.txt
coverage==4.0.2
Django==1.8.5
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.5.2
selenium==2.48.0
